Rename getSymbolTable.ts to symbols.ts
