+++ src/common/symbols.ts
